replace scanner.py w/ curses implementation
